Encapsulate single and dual versions of graphs as (functional) components
