+++ src/components/graphs/SingleTimeSeriesGraph.js
@@ -1,0 +1,41 @@
+import React from 'react';
+
+import _ from 'underscore';
+
+import { timeseriesToTimeseriesGraph } from '../../core/chart';
+import TimeSeriesGraph from './TimeSeriesGraph';
+
+
+export default function SingleTimeSeriesGraph(props) {
+  function getMetadata() {
+    const {
+      model_id, experiment,
+      variable_id, meta,
+    } = props;
+
+    const primaryVariableMetadata = _.findWhere(meta, {
+      model_id, experiment, variable_id,
+    });
+    // Yes, the value of this function is an array of one element.
+    const metadataSets = [primaryVariableMetadata];
+    return metadataSets;
+  }
+
+  function dataToGraphSpec(meta, data) {
+    // Convert `data` (described by `meta`) to a graph specification compatible
+    // with `DataGraph`.
+    return timeseriesToTimeseriesGraph(meta, ...data);
+  }
+
+  const graphProps = _.pick(props,
+    'model_id', 'variable_id', 'experiment', 'meta', 'area'
+  );
+
+  return (
+    <TimeSeriesGraph
+      {...graphProps}
+      getMetadata={getMetadata}
+      dataToGraphSpec={dataToGraphSpec}
+    />
+  );
+}
